+++ src/main.rs
@@ -1,14 +1,5 @@
-use std::time::Instant;
-use chrono::{DateTime, NaiveDateTime};
 use log::info;
-use crate::traits::AsDataFrame;
-use crate::backtesting::{BacktestingConfig, BacktestingRunner};
-use crate::portfolio::PortfolioArgs;
-
-use polars::prelude::*;
-use rust_decimal_macros::dec;
-use crate::markets::utils::save_candles;
-use crate::utils::print_candle_statistics;
+use crate::backtesting::BacktestingRuntime;
 
 mod backtesting;
 mod indicators;
@@ -16,7 +7,6 @@ mod markets;
 mod portfolio;
 mod serialization;
 mod strategies;
-mod timing;
 mod traits;
 mod types;
 mod utils;
@@ -27,54 +17,16 @@ mod processor;
 fn main() {
     colog::init();
 
-    let db_path = "data/candle_data.sqlite3";
-    let candle_table = "coinbase_SHIBUSD_5m_candles";
-    let market_data_table = "coinbase_BTCUSD_5m_candles";
-
-    let candles = utils::extract_candles_from_db(db_path, candle_table).unwrap().as_dataframe();
-    let market_data = utils::extract_candles_from_db(db_path, market_data_table).unwrap().as_dataframe();
-
     let strategy = strategies::Strategy::new(
         vec![Box::new(indicators::BBands::default())],
         strategies::Consensus::Unison,
     );
 
-    let mut runner = BacktestingRunner::from_config(
+    let mut runtime = BacktestingRuntime::from_config(
         "data/backtesting_config.toml",
         strategy
-    );
+    ).load_candles().expect("Could not load candles");
 
-    let start_time = Instant::now();
     info!("Starting to process");
-    let performance = runner.run(&candles, &market_data).unwrap();
-    let elapsed = start_time.elapsed();
-
-    print_candle_statistics(&candles);
-
-    let candle_len = candles.height();
-    println!("Finished processing {:?} rows in {:?}", candle_len, elapsed);
-    println!("Avg. processing time per row: {:?}", elapsed / candle_len as u32);
-
-    println!("Performance: {:?}", performance);
-
-
-    // save candles and indicator graphs as CSV
-
-    info!("Saving data to CSV");
-
-    let save_path = "data";
-
-    let candle_path = format!("data/{}.csv", candle_table);
-    let market_data_path = format!("data/{}.csv", market_data_table);
-
-    // create mutable versions of the DataFrames
-    let mut candles = candles.clone();
-    let mut market_data = market_data.clone();
-
-    // save candles to CSV
-    save_candles(&mut candles, &candle_path).unwrap();
-    save_candles(&mut market_data, &market_data_path).unwrap();
-
-    // save indicator graph as CSV
-    runner.save_indicator_data(save_path).unwrap()
+    runtime.run().unwrap();
 }